WIP bundle tests
